% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redq-package.R
\docType{package}
\name{redq-package}
\alias{redq}
\alias{redq-package}
\title{redq: empirical dynamic quantiles for time series}
\description{
A semi-vectorised parallel-looped algorithm for computing empirical dynamic quantiles for large temporal data sets.
}
\author{
\strong{Maintainer}: Stephen Su \email{szmsu2011@hotmail.co.nz} (\href{https://orcid.org/0009-0002-5402-0667}{ORCID})

}
